Create index.html, style.css, and add ref images
This is a project to create a landing page for a website.

I created the basis for the html and css of the project. Imported the
Google Roboto font (which took me such long time). I also added
the reference images to the project.

This is the beginning. The beginning. This is the beginning of
HTML5.